backend/requirements.txt


fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
playwright==1.44.0
neo4j
google-generativeai
langchain-google-genai